Rename hdf52 to hdf5ea, basic functionality works now.
Use PyTables EArrays instead of Tables so that large data vectors work.
